+++ hahomematic/platforms/entity.py
@@ -299,6 +299,7 @@ class BaseEntity(CallbackEntity, PayloadMixin):
 
         self._forced_usage: EntityUsage | None = None
         self._entity_name_data: Final = self._get_entity_name()
+        self._collector_methods: dict[str, Callable] = {}
 
     @state_property
     def available(self) -> bool:
@@ -309,6 +310,17 @@ class BaseEntity(CallbackEntity, PayloadMixin):
     def base_channel_no(self) -> int | None:
         """Return the base channel no of the entity."""
         return self._device.get_sub_device_channel(channel_no=self._channel_no)
+
+    # @property
+    @property
+    def collector_methods(self) -> Mapping[str, Callable]:
+        """Return all collector methods."""
+        return self._collector_methods
+
+    @property
+    def collector_method_names(self) -> tuple[str, ...]:
+        """Return all collector methods."""
+        return tuple(self._collector_methods.keys())
 
     @property
     def path(self) -> str:
@@ -446,6 +458,7 @@ class BaseParameterEntity[
         self._state_uncertain: bool = True
         self._is_forced_sensor: bool = False
         self._assign_parameter_data(parameter_data=parameter_data)
+        self._collector_methods = get_bind_collector_methods(obj=self)
 
     def _assign_parameter_data(self, parameter_data: ParameterData) -> None:
         """Assign parameter data to instance variables."""
@@ -855,7 +868,6 @@ def bind_collector(
             """Wrap method to add collector."""
             if not enabled:
                 return await func(*args, **kwargs)
-
             try:
                 collector_exists = args[argument_index] is not None
             except IndexError:
@@ -874,6 +886,19 @@ def bind_collector(
                 )
             return return_value
 
+        wrapper_collector.bind_collector = True  # type: ignore[attr-defined]
         return wrapper_collector  # type: ignore[return-value]
 
     return decorator_bind_collector
+
+
+def get_bind_collector_methods(obj: object) -> dict[str, Callable]:
+    """Get all methods decorated with the "bind_collector" decorator."""
+    return {
+        name: getattr(obj, name)
+        for name in dir(obj)
+        if not name.startswith("_")
+        # and name != "collector_methods"
+        and callable(getattr(obj, name))
+        and hasattr(getattr(obj, name), "bind_collector")
+    }